requirements.txt
spotipy==2.23.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Pillow==10.1.0
lxml==4.9.3
customtkinter==5.2.0

